Implement save bookmark window as iFrame in popup
Implements the save bookmark window as an iFrame in the popup.html window.
Additionally, also passes selected text from the page being bookmarked as
the description field of the bookmark.

NOTE: Have not updated the CRX file, as I do not have the private key.

Having the bookmark window as part of the popup is purely personal
preference (aesthetics), there are some downsides (to-dos):

* If you half enter some details, but then click off the popup without
  saving the bookmark, the text will be lost. This is probably fixable
  by making the extenstion remember the last state.
* Can't lanuch mutliple windows for multiple tabs (perhaps with the
  intention of saving bookmarks later - I wouldn't have thought anyone
  would work this way, but you never know). Probably can't fix this.
* If using the pulldown for tag auto-complete there can be a ugly scroll
  bar visible in the popup, until the pulldown list reduces in size.
  Perhaps add a bit of 'padding' onto the popup or prevent overflow to
  solve this.