Use local storage to draft bookmark tags and descriptions
Since forking this extension and implementing the Pinboard add bookmark
window as an iFrame within the extension popup, this has had the one
downside that if you enter tags and description, but click outside of the
extension popup before saving the bookmark any text you enter is lost.
Quite annoying, as often you can refer back to the page or other pages when
coming up with tags and descriptions.

This extension now uses HTML5 Local Storage to save and restore tags and
descriptions so you don't lose text if you accidentally (or on purpose)
close the popup before saving the bookmark.

Still work to do though. Caveats:

 - works globally so you can only "draft" one bookmark, i.e. can't
   draft tags and description for different tabs simultaneously.
 - Is a little bit clever though: if you are drafting a bookmark
   from tab A and close the popup and switch to tab B and click on
   the save bookmark option it won't fill in the draft values from
   tab A (since it is a different url). It also won't overwrite the
   the stored values unless you enter some info whilst on tab B. I.e.
   that means you can close the popup (say you got the wrong tab),
   switch back to tab A and the draft values will still be there.
 - will attempt to restore values only if the fields are blank. i.e.
   an unsaved bookmark. Also, no good for drafting changes to previously
   saved bookmarks.
 - Only works on the tags and description fields.
 - Effects the pinboard.in bookmarklets and any window saving via
   the "https://pinboard.in/add/" url.
 - Should probably be using messaging bewteen iFrame and extension to do
   local storage from extension rather than directly in the iFrame.