Bug fix: Enable save bookmark without tab reload on 1st install of ext
Always declare your variables! My bad mistake.

Since the ext uses a content script to get the selected text and use it for
the bookmark description, it would throw an error on first install if you
tried to save a bookmark for an existing tab, because obviously the content
script had not existed for that tab previously. This is trivial to solve
simply by declaring the damn variable in the first place. Oh deary me.

Sorry!