Start implementing Read Later in the background
This commit just for record purposes. I wanted to implement the Read
Later functionality without opening up a popup window. Had thought about
doing it in an iframe as per save bookmark, but then decided to do it in
a background page.

Note that the icon changing isn't working yet. I need to move this code
to the background page as I think the window.close() is ending
everything prematurely. But wanted to record this attempt for reference.