Move add/read later functionality to background page
As mentioned in previous commit. Move this function to background page so
can close the popup immediately and still allow function to run (especially
the icon changing bit; previously the window.close() was preventing the
icon from changing back).

Lots more work still required such as changing icon colour only if
successfully saved, etc. Checking if logged in, etc. But this is a good
enough stab at it for now.